Reworked home to use folders
