+++ web/app/integration-manager/_merlins-connector/home/commands.js
@@ -1,0 +1,11 @@
+import {fetchPageData} from '../utils'
+import homeParser from './parser'
+import {receiveHomeData} from '../../responses'
+
+export const fetchHomeData = (url) => (dispatch) => {
+    return dispatch(fetchPageData(url))
+        .then(([$, $response]) => {
+            dispatch(receiveHomeData(homeParser($, $response)))
+        })
+}
+
