+++ web/app/integration-manager/commands.js
@@ -1,7 +1,11 @@
+import {register as registerHome} from './home/commands'
+
 let connector = {}
 
 export const register = (commands) => {
     connector = commands
+
+    registerHome(commands.home)
 }
 
 export const fetchPdpData = (...args) => connector.fetchPdpData(...args)
@@ -15,5 +19,3 @@ export const submitShipping = (...args) => connector.submitShipping(...args)
 export const checkCustomerEmail = (...args) => connector.checkCustomerEmail(...args)
 
 export const submitSignIn = (...args) => connector.submitSignIn(...args)
-
-export const fetchHomeData = (...args) => connector.fetchHomeData(...args)
