+++ web/app/integration-manager/home/commands.js
@@ -1,0 +1,7 @@
+let connector = {}
+
+export const register = (commands) => {
+    connector = commands
+}
+
+export const fetchHomeData = (...args) => connector.fetchHomeData(...args)
